#301 combine zocalo callback and interaction
